% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_playground.R
\name{create.enum.simple}
\alias{create.enum.simple}
\title{Make enums with assigned values or use the name as the value if no value is specified}
\usage{
create.enum.simple(inputList)
}
\arguments{
\item{inputList}{}
}
\description{
Defines the enum by constructing a list like list(a = "a", b = "b", ...)
}
