% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_as_data_frame.R
\name{enum.as.data.frame}
\alias{enum.as.data.frame}
\title{Creates a data.frame with all elements of an enum}
\usage{
enum.as.data.frame(enum)
}
\arguments{
\item{enum}{The enumeration (list of constants)}
}
\value{
A data.frame with three columns: allowed.values, value.names and descriptions.
              The 'descriptions' column is set to the names attribute of the enum if no 'descriptions'
              attribute exists.
}
\description{
The data.frame is useful for debugging purposes, filling list and combo boxes in user interfaces to
select one value etc.
}
