% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_enumeration.R
\name{is.enumeration}
\alias{is.enumeration}
\title{Test if an object is an instance of the class \code{enumeration} (created with \code{\link{create.enum}}}
\usage{
is.enumeration(obj)
}
\arguments{
\item{obj}{object to be tested}
}
\value{
TRUE if \code{obj} is an instance of the class \code{enumeration}
}
\description{
Test if an object is an instance of the class \code{enumeration} (created with \code{\link{create.enum}}
}
\examples{
is.enumeration(1.10)
is.enumeration(create.enum(1:10))
}
