% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_package_doc.R
\docType{package}
\name{R_enumerations}
\alias{R_enumerations}
\alias{R_enumerations-package}
\title{Adds an enumeration data type for R to support list of allowed value/label pairs}
\description{
Enumerations are lists of value/label pairs (lists of constants for values).
Enums make code more readable.
You can use an enumeration type as formal function parameter to allow only values
of the enumeration passed in (other values throw an error).
}
