% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data_frame_enumeration.R
\name{as.data.frame.enumeration}
\alias{as.data.frame.enumeration}
\title{Creates a data.frame with all elements of an enum}
\usage{
\method{as.data.frame}{enumeration}(enum)
}
\arguments{
\item{enum}{An object of class "enumeration" (created with \code{\link{create.enum}})}
}
\value{
A data.frame with three columns: allowed.values, value.names and descriptions.
              The 'descriptions' column is set to the names attribute of the enum if no 'descriptions'
              attribute exists.
}
\description{
The data.frame is useful for debugging purposes, filling list and combo boxes in user interfaces to
select one value etc.
}
