% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_playground.R
\name{create.enum}
\alias{create.enum}
\title{Create an enumeration}
\usage{
create.enum(allowed.values, value.names = make.names(allowed.values, unique =
  TRUE))
}
\arguments{
\item{allowed.values}{}

\item{value.names}{}
}
\description{
Create an enumeration
}
