% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_enum.R
\name{create.enum}
\alias{create.enum}
\title{Create an enumeration (an enumeration is a list of constants)}
\usage{
create.enum(allowed.values, value.names = allowed.values,
  descriptions = value.names, ensure.valid.value.names = TRUE)
}
\arguments{
\item{allowed.values}{Vector with all allowed values}

\item{value.names}{Vector of character containing the names that correspond to the allowed values
Duplicated names will be made unique.
Invalid characters are replaced by characters that are allowed in names in R}

\item{descriptions}{Vector with more detailled descriptive information for each enum value}

\item{ensure.valid.value.names}{TRUE to convert invalid characters into syntacically allowed names
     and make duplicated names unique (by appending a number).
FALSE leaves value names like they are. To use enum names that contain invalid
      characters you have to quote them, e. g. \code{my.enum$`a special name`}}
}
\value{
An object of class "enumeration" that represents an enumeration via a list with named elements
}
\description{
This helper function builds an enum type using the provided arguments. It is mainly useful
to load the elements of an enum from a data base or config file (e. g. CSV file) and create an enum type.
}
\examples{
               
COLOR.ENUM <- create.enum(c(1L, 2L, 3L), c("BLUE", "RED", "BLACK"))
# Returns an enumeration type that also could have been constructed manually like this:
COLOR.ENUM <- list(BLUE = 1L, RED = 2L, BLACK = 3L)
}
