% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_enum_arg.R
\name{match.enum.arg}
\alias{match.enum.arg}
\title{Match the passed enum value against the list of allowed enum values}
\usage{
match.enum.arg(arg, choices)
}
\arguments{
\item{arg}{}

\item{choices}{}
}
\description{
Based on the R code of the base function "match.arg"...
}
